+++ lib/States/include/StateMachine.h
@@ -2,9 +2,9 @@
 #define STATES_STATE_MACHINE_H
 
 
-#include "RelayControl.h"
-#include "StatusHandler.h"
-#include "PressureSensor.h"
+#include <RelayControl.h>
+#include <StatusHandler.h>
+#include <PressureSensor.h>
 #include <State.h>
 #include <StatePumpOff.h>
 #include <StatePumpOn.h>
