+++ lib/States/include/StatePumpOff.h
@@ -1,7 +1,7 @@
 
 #ifndef STATE_PUMP_OFF_H
 #define STATE_PUMP_OFF_H
-#include "State.h"
+#include <State.h>
 
 class StatePumpOff : public State
 {
