+++ lib/States/include/StatePumpOn.h
@@ -1,7 +1,7 @@
 
 #ifndef STATE_PUMP_ON_H
 #define STATE_PUMP_ON_H
-#include "State.h"
+#include <State.h>
 
 class StatePumpOn : public State
 {
