+++ platformio.ini
@@ -9,7 +9,7 @@
 ; https://docs.platformio.org/page/projectconf.html
 
 [env:native]
-test_framework = unity
+test_framework = googletest
 platform = native
 build_flags = -DUNIT_TEST
 lib_extra_dirs = ~/Documents/Arduino/libraries
@@ -17,14 +17,14 @@ lib_extra_dirs = ~/Documents/Arduino/libraries
 ; 	-<well-pump-control.cpp>
 lib_deps = 
 	fabiobatsilva/ArduinoFake@^0.4.0
+	wollewald/ADS1115_WE@^1.5.1
 	marcoschwartz/LiquidCrystal_I2C@^1.1.4
 	arduino-libraries/ArduinoBLE@^1.3.6
 	powerbroker2/SafeString@^4.1.33
 	robtillaart/MAX6675@^0.3.2
-	throwtheswitch/Unity@^2.5.2
-
+	
 [env:nano_33_iot]
-test_framework=unity
+test_framework=googletest
 test_ignore = native_tests/*
 platform = atmelsam
 board = nano_33_iot
