+++ src/well-pump-control.cpp
@@ -11,9 +11,9 @@
 // install SafeString library from Library manager or from https://www.forward.com.au/pfod/ArduinoProgramming/SafeString/index.html
 // to get BufferedOutput. See https://www.forward.com.au/pfod/ArduinoProgramming/Serial_IO/index.html for a full tutorial
 // on Arduino Serial I/O that Works
-#include "RelayControl.h"
-#include "StatusHandler.h"
-#include "PressureSensor.h"
+#include <RelayControl.h>
+#include <StatusHandler.h>
+#include <PressureSensor.h>
 #include <StateMachine.h>
 
 /*
